feat: Add lighthouse rendering for ghosts & coins
Add rendering of lighthouse colors for ghost entities on a higher z level to
enhance visualization. Update lighthouse position based on ghost translation
values.

+++ src/map/mod.rs
@@ -32,7 +32,7 @@ fn spawn_tiles(mut commands: Commands, map: Res<TileMap>, ascii: Res<AsciiSheet>
                 Tile::Wall => {
                     spawn_sprites_for_wall(&mut commands, &ascii, &map, x as i32, y as i32);
                 }
-                Tile::Coin => spawn_coin(&mut commands, &ascii, x as i32, y as i32),
+                Tile::Coin => spawn_coin(&mut commands, &ascii, x, y),
                 _ => {
                     continue;
                 }
